Add CHData to decide which src to use
- Add CHData service for choosing whether to access CSV/Hive based on env variable 'source'
- Update README.md tables
- Remove unused code in SearchController
- Remove unused routes
